services/operator_console/requirements_local.txt
kafka-python==2.0.2
flask==3.0.0
flask-cors==4.0.0
